Refactor by splitting out files
