+++ requirements.txt
@@ -1,9 +1,9 @@
-Flask==0.12.1 \
-    --hash=sha256:6c3130c8927109a08225993e4e503de4ac4f2678678ae211b33b519c622a7242 \
-    --hash=sha256:9dce4b6bfbb5b062181d3f7da8f727ff70c1156cbb4024351eafd426deb5fb88
-Flask-SQLAlchemy==2.2 \
-    --hash=sha256:f0d8241efba723d7b878f73550f5d3c0fbb042416123b52b36640b7491fa208b \
-    --hash=sha256:fa3bd6343de231b2b3376a276df6d331c63c766449b660fc04d375a1c21312ac
+Flask==0.12.4 \
+    --hash=sha256:6c02dbaa5a9ef790d8219bdced392e2d549c10cd5a5ba4b6aa65126b2271af29 \
+    --hash=sha256:2ea22336f6d388b4b242bc3abf8a01244a8aa3e236e7407469ef78c16ba355dd
+Flask-SQLAlchemy==2.3.2 \
+    --hash=sha256:3bc0fac969dd8c0ace01b32060f0c729565293302f0c4269beed154b46bec50b \
+    --hash=sha256:5971b9852b5888655f11db634e87725a9031e170f37c0ce7851cf83497f56e53
 gunicorn==19.6.0 \
     --hash=sha256:723234ea1fa8dff370ab69830ba8bc37469a7cba13fd66055faeef24085e6530 \
     --hash=sha256:813f6916d18a4c8e90efde72f419308b357692f81333cb1125f80013d22fb618
@@ -71,9 +71,9 @@ cryptography==1.7.1 \
     --hash=sha256:cf5af01fe7b471745a0266679565c640df3e151e366bd57584997e10acf8454c \
     --hash=sha256:c087a345510e56f944298e76b84c0c86b0aaaee3cd29fb1112d5077de025f484 \
     --hash=sha256:953fef7d40a49a795f4d955c5ce4338abcec5dea822ed0414ed30348303fdb4c
-requests==2.12.4 \
-    --hash=sha256:000748df49e087784441b2621c50fb81046c5c8e80e0d91674ffad65b9e13844 \
-    --hash=sha256:ed98431a0631e309bb4b63c81d561c1654822cb103de1ac7b47e45c26be7ae34
+requests==2.21.0 \
+    --hash=sha256:7bf2a778576d825600030a110f3c0e3e8edc51dfaafe1c146e39a2027784957b \
+    --hash=sha256:502a824f31acdacb3a35b6690b5fbf0bc41d63a24a45c4004352b0242707598e
 click==2.0 \
     --hash=sha256:d2752f89b1ebba748873c19662c2b2bfeccf8171e316fce0e1fb56066ceae178 \
     --hash=sha256:33a984fe73aa8c6bf69c38098b69dd38405529db54baf4c8cf64c50b2dee644f
@@ -84,8 +84,8 @@ Jinja2==2.4 \
     --hash=sha256:767418380bf7ec1701414f9f361ca6c261f705dbb576f52366f7d5ebc1b25f1d
 itsdangerous==0.21 \
     --hash=sha256:c760381fc09dc120e4b459f5ba39ed55ea21fd77ac9f4d4101c0c237d8fc9d9e
-SQLAlchemy==1.1.13 \
-    --hash=sha256:2a98ac87b30eaa2bee1f1044848b9590e476e7f93d033c6542e60b993a5cf898
+SQLAlchemy==1.2.16 \
+    --hash=sha256:6af3ca2f7f00844465ab4fa78337d487b39e53f516c51328aed4ed3a719d4264
 six==1.10.0 \
     --hash=sha256:0ff78c403d9bccf5a425a6d31a12aa6b47f1c21ca4dc2573a7e2f32a97335eb1 \
     --hash=sha256:105f8d68616f8248e24bf0e9372ef04d3cc10104f1980f54d57b2ce73a5ad56a
@@ -121,9 +121,9 @@ cffi==1.11.2 \
     --hash=sha256:5357b465e3d6b98972b7810f9969c913d365e75b09b7ba813f5f0577fe1ac9f4 \
     --hash=sha256:75e1de9ba7c155d89bcf67d149b1c741df553c8158536e8d27e63167403159af \
     --hash=sha256:ab87dd91c0c4073758d07334c1e5f712ce8fe48f007b86f8238773963ee700a6
-idna==2.0 \
-    --hash=sha256:9b2fc50bd3c4ba306b9651b69411ef22026d4d8335b93afc2214cef1246ce707 \
-    --hash=sha256:16199aad938b290f5be1057c0e1efc6546229391c23cea61ca940c115f7d3d3b
+idna==2.8 \
+    --hash=sha256:ea8b7f6188e6fa117537c3df7da9fc686d485087abf6ac197f9c46432f7e4a3c \
+    --hash=sha256:c357b3f628cf53ae2c4c05627ecc484553142ca23264e593d327bcde5e9c3407
 pyasn1==0.1.8 \
     --hash=sha256:b232366d08266df66dd7eac3309543f0287a0d3a27dd7661cf32a299cbd7c401 \
     --hash=sha256:7dd74178fae71a1e39d57f2b1a87bc3779834c69ed0477ca80f57bf901981c50 \
@@ -146,3 +146,12 @@ jdcal==1.4 \
     --hash=sha256:ea0a5067c5f0f50ad4c7bdc80abad3d976604f6fb026b0b3a17a9d84bb9046c9
 et_xmlfile==1.0.1 \
     --hash=sha256:614d9722d572f6246302c4491846d2c393c199cfa4edc9af593437691683335b
+certifi==2018.11.29 \
+    --hash=sha256:993f830721089fef441cdfeb4b2c8c9df86f0c63239f06bd025a76a7daddb033 \
+    --hash=sha256:47f9c83ef4c0c621eaef743f133f09fa8a74a9b75f037e8624f83bd1b6626cb7
+chardet==3.0.4 \
+    --hash=sha256:fc323ffcaeaed0e0a02bf4d117757b98aed530d9ed4531e3e15460124c106691 \
+    --hash=sha256:84ab92ed1c4d4f16916e05906b6b75a6c0fb5db821cc65e70cbd64a3e2a5eaae
+urllib3==1.24.1 \
+    --hash=sha256:61bf29cada3fc2fbefad4fdf059ea4bd1b4a86d2b6d15e1c7c0b582b9752fe39 \
+    --hash=sha256:de9529817c93f27c8ccbfead6985011db27bd0ddfcdb2d86f3f663385c6a9c22
