+++ test_requirements.txt
@@ -1,5 +1,5 @@
-flake8==3.3.0
+flake8==3.6.0
 mock==2.0.0
 testing.postgresql==1.3.0
-pytest==3.2.3
-pytest-cov==2.5.1
+pytest==4.1.1
+pytest-cov==2.6.1
